common: Fix hardcoded Python paths
It is important that python scripts not hardcode a specific python3
path. Even if the python at that path is a suitable version, there is
no guarantee that the packages it has installed at that path are
correct.

It is also considered "bad form" to treat the system python directory
as fair game for modifications. The idiomatic method of handling this
in the python world is to use a separate environment, or even a
"python virtual environment".

For now, we should just assume that whatever python is in the user's
path is the python they want us to use.

```
$ make
Sort failed: module 'networkx' has no attribute 'topological_generations'
make: *** [Makefile:2: all] Error 1
$
```